Refactored Channel's c code variable declaration for 32-bit 6-4bit compatible.

+++ native/ax/Channel-2/ChannelDataMng.c
@@ -75,8 +75,8 @@
 #include <stdarg.h>
 
 #include <linux/dvb/frontend.h>
-int TranslateJWideString2( char* buf_out, int buf_out_size, unsigned short* buf_in );
-int TranslateJWideString( char* buf_out, int buf_out_size, unsigned short* buf_in );
+int TranslateJWideString2( char* buf_out, int buf_out_size, uint16_t* buf_in );
+int TranslateJWideString( char* buf_out, int buf_out_size, uint16_t* buf_in );
 
 
 //*********************** LINUX section ***********************
@@ -181,7 +181,7 @@ void _set_ChannelDataMng_log( bool enable )
 */
 static int  getNextTextField( char* p_in, char **p_out, char* text, int max_len );
 static bool getNextIntField( char* p_in, char **p_out, int* int_val );
-static bool getNextShortField( char* p_in, char **p_out, unsigned short* short_val );
+static bool getNextShortField( char* p_in, char **p_out, uint16_t* short_val );
 static bool parseATSCFreqEntry( char* p_in, ATSC_FREQ* atsc_freq ) ;
 static bool parseQAMFreqEntry( char* p_in, QAM_FREQ* qam_freq );
 static bool parseUserQAMFreqData( char* p_in, struct bcast_qam *Freq, char **error );
@@ -196,7 +196,7 @@ char* makeDVBCFreqString( char*Buf, int Size,  DVB_C_FREQ * Freq  );
 static bool  parseDVBSFreqEntry( char* p_in, DVB_S_FREQ * Freq, char** error  );
 char* makeDVBSFreqString( char*Buf, int Size,  DVB_S_FREQ * Freq  );
 bool GetString( char*p, const char* Name, char* Str, int MaxLen );
-unsigned long fileTimeStamp( char* szFileName );
+uint32_t fileTimeStamp( char* szFileName );
 static char version_string[] ="VERSION 3.0, country:%s\r\n";
 static char type_string[] ="TYPE %s\r\n";
  
@@ -589,7 +589,7 @@ int saveScanTuningTable( CHANNEL_DATA *Channel, char* file_ext )
 				fprintf( fp, "%s\r\n", makeATSCFreqString( buf, sizeof(buf),  (ATSC_FREQ *)&ATSCFreq[i]  ) );
 			} 
 			/*else
-				fprintf( fp, "ch:%d %d %d %d %d %ld %ld #%s\r\n",
+				fprintf( fp, "ch:%d %d %d %d %d %d %d #%s\r\n",
 			          ATSCFreq[i].index,   ATSCFreq[i].major,       ATSCFreq[i].minor,
 				      ATSCFreq[i].program, ATSCFreq[i].physical_ch, ATSCFreq[i].frequency, 
 				      ATSCFreq[i].ctrl,    ATSCFreq[i].name );	
@@ -623,7 +623,7 @@ int saveScanTuningTable( CHANNEL_DATA *Channel, char* file_ext )
 			}
 			/*
 			else
-				fprintf( fp, "ch:%d %d %d %d %d %ld %d %d %ld #%s\r\n",
+				fprintf( fp, "ch:%d %d %d %d %d %d %d %d %d #%s\r\n",
 			          QAMFreq[i].index,   QAMFreq[i].major,       QAMFreq[i].minor,
 				      QAMFreq[i].program, QAMFreq[i].physical_ch, QAMFreq[i].frequency,
 				      QAMFreq[i].modulation, QAMFreq[i].inversal, 
@@ -1160,7 +1160,7 @@ int saveCacheQAMFreq(  CHANNEL_DATA *Channel, int freqTableType, struct bcast_qa
 	i = 0;
 	while ( freqTbl[i].fq != 0 && i<256 ) 
 	{
-		fprintf( fp, "CH:%d frq:%ld \r\n", freqTbl[i].ch, freqTbl[i].fq );
+		fprintf( fp, "CH:%d frq:%d \r\n", freqTbl[i].ch, freqTbl[i].fq );
 		i++;
 	}
 	fprintf( fp, "CH:%d frq:0000 \r\n", 999 );
@@ -1331,13 +1331,13 @@ static bool getNextIntField( char* p_in, char **p_out, int* int_val )
 	return true;
 }
 
-static bool getNextShortField( char* p_in, char **p_out, unsigned short* short_val )
+static bool getNextShortField( char* p_in, char **p_out, uint16_t* short_val )
 {
 	char *token, *p;
 	token = (char*)strtok_r( p_in, delima, &p );
 	if ( token == NULL ) 
 		return false;
-	*short_val = (unsigned short)atoi( token );
+	*short_val = (uint16_t)atoi( token );
 	*p_out = p;
 	return true;
 }
@@ -1681,7 +1681,7 @@ static bool GetExtParameter( char* p, char* Parameter, int MaxLen  )
 }
 
 
-bool GetIntVal( char*p, const char* Name, unsigned long* pVal )
+bool GetIntVal( char*p, const char* Name, uint32_t* pVal )
 {
 	char *s, *e;
 	e = p;
@@ -1710,7 +1710,7 @@ bool GetIntVal( char*p, const char* Name, unsigned long* pVal )
 	return false;
 }
 
-bool GetWordVal( char*p, const char* Name, unsigned short* pVal )
+bool GetWordVal( char*p, const char* Name, uint16_t* pVal )
 {
 	char *s, *e;
 	e = p;
@@ -1729,7 +1729,7 @@ bool GetWordVal( char*p, const char* Name, unsigned short* pVal )
 				while( *s && ( *s == ' ' || *s == '\t' ) ) s++; //skip white space
 				if ( *s )
 				{
-					*pVal = (unsigned short)atoi( s );
+					*pVal = (uint16_t)atoi( s );
 					return true;
 				}
 			}
@@ -1819,7 +1819,7 @@ bool GetQuoteString( char*p, const char* Name, char* Str, int MaxLen )
 static bool parseDVBTFreqEntry( char* p, DVB_T_FREQ * Freq, char** error  )
 {
 	bool found = true;
-	unsigned long val;
+	uint32_t val;
 	if ( Freq == NULL )
 		return false;
 
@@ -1894,7 +1894,7 @@ static bool parseDVBTFreqEntry( char* p, DVB_T_FREQ * Freq, char** error  )
 static bool parseDVBCFreqEntry( char* p, DVB_C_FREQ * Freq, char** error  )
 {
 	bool found = true;
-	unsigned long val;
+	uint32_t val;
 	if ( Freq == NULL )
 		return false;
 
@@ -1956,7 +1956,7 @@ static bool parseDVBCFreqEntry( char* p, DVB_C_FREQ * Freq, char** error  )
 static bool parseDVBSFreqEntry( char* p, DVB_S_FREQ * Freq, char** error  )
 {
 	bool found = true;
-	unsigned long val;
+	uint32_t val;
 	if ( Freq == NULL )
 		return false;
 
@@ -2046,9 +2046,9 @@ static bool parseDVBSFreqEntry( char* p, DVB_S_FREQ * Freq, char** error  )
 static bool parseUserQAMFreqData( char* p, struct bcast_qam *Freq, char **error )
 {
 	bool found = true;
-	unsigned long val;
+	uint32_t val;
 	int frq;
-	unsigned short ch;
+	uint16_t ch;
 	if ( Freq == NULL )
 		return false;
 
@@ -2056,13 +2056,13 @@ static bool parseUserQAMFreqData( char* p, struct bcast_qam *Freq, char **error 
 	
 	Freq->fq = 0; 	Freq->ch = 0;  Freq->inv = 0;  Freq->mod = 0; 
 	if ( !GetWordVal( p, "CH",   &ch) && !GetWordVal( p, "ch",   &ch ) )  { found = false; if ( error && *error == 0x0 ) *error = "index not found 'CH:xx'"; }
-	if ( !GetIntVal( p, "frq",   (unsigned long*)&frq) ) { found = false; if ( error && *error == 0x0 ) *error = "frequency not found"; }
+	if ( !GetIntVal( p, "frq",   (uint32_t*)&frq) ) { found = false; if ( error && *error == 0x0 ) *error = "frequency not found"; }
 
 	if ( !found )
 		return found;
 
 	Freq->ch = (int)ch;
-	Freq->fq = (unsigned long)frq;
+	Freq->fq = (uint32_t)frq;
 	if ( GetIntVal( p, "mod", &val ) )
 		Freq->mod = (char)val;
 
@@ -2075,8 +2075,8 @@ static bool parseUserQAMFreqData( char* p, struct bcast_qam *Freq, char **error 
 static bool parseScanATSCFreqData( char* p, ATSC_FREQ *Freq, char **error )
 {
 	bool found = false;
-	unsigned long val;
-	unsigned short n;
+	uint32_t val;
+	uint16_t n;
 	if ( Freq == NULL )
 		return false;
 
@@ -2102,10 +2102,10 @@ static bool parseScanATSCFreqData( char* p, ATSC_FREQ *Freq, char **error )
 	}
 
 	if ( GetIntVal( p, "frq", &val) ) 
-		Freq->frequency = (unsigned long)val;
+		Freq->frequency = (uint32_t)val;
 
 	if ( GetIntVal( p, "frq", &val) ) 
-		Freq->frequency = (unsigned long)val;
+		Freq->frequency = (uint32_t)val;
 
 	if ( GetIntVal( p, "v_pid", &val ) )
 		Freq->video_pid = (char)val;
@@ -2126,8 +2126,8 @@ static bool parseScanATSCFreqData( char* p, ATSC_FREQ *Freq, char **error )
 static bool parseScanQAMFreqData( char* p, QAM_FREQ *Freq, char **error )
 {
 	bool found = false;
-	unsigned long val;
-	unsigned short n;
+	uint32_t val;
+	uint16_t n;
 	if ( Freq == NULL )
 		return false;
 
@@ -2153,7 +2153,7 @@ static bool parseScanQAMFreqData( char* p, QAM_FREQ *Freq, char **error )
 	}
 
 	if ( GetIntVal( p, "frq", &val) ) 
-		Freq->frequency = (unsigned long)val;
+		Freq->frequency = (uint32_t)val;
 
 	if ( GetIntVal( p, "mod", &val ) )
 		Freq->modulation = (char)val;
@@ -2194,7 +2194,7 @@ char* makeATSCFreqString( char*Buf, int Size,  ATSC_FREQ * Freq  )
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	SPRINTF( tmp, sizeof(tmp), "phy:%d ", Freq->physical_ch );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "frq:%ld ", Freq->frequency );
+	SPRINTF( tmp, sizeof(tmp), "frq:%d ", Freq->frequency );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 
 	if ( Freq->name[0] )
@@ -2227,7 +2227,7 @@ char* makeQAMFreqString( char*Buf, int Size,  QAM_FREQ * Freq  )
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	SPRINTF( tmp, sizeof(tmp), "phy:%d ", Freq->physical_ch );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "frq:%ld ", Freq->frequency );
+	SPRINTF( tmp, sizeof(tmp), "frq:%d ", Freq->frequency );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	SPRINTF( tmp, sizeof(tmp), "mod:%d ", Freq->modulation );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
@@ -2260,9 +2260,9 @@ char* makeDVBTFreqString( char*Buf, int Size,  DVB_T_FREQ * Freq  )
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	SPRINTF( tmp, sizeof(tmp), "sid:%d ", Freq->sid );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "frq:%ld ", Freq->frequency );
+	SPRINTF( tmp, sizeof(tmp), "frq:%d ", Freq->frequency );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "band:%ld ", Freq->bandwidth );
+	SPRINTF( tmp, sizeof(tmp), "band:%d ", Freq->bandwidth );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	
 	if ( Freq->trans_mode != 0xff ) 
@@ -2303,7 +2303,7 @@ char* makeDVBTFreqString( char*Buf, int Size,  DVB_T_FREQ * Freq  )
 	}
 
 
-	SPRINTF( tmp, sizeof(tmp), "ctrl:%ld ", Freq->ctrl );
+	SPRINTF( tmp, sizeof(tmp), "ctrl:%d ", Freq->ctrl );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	
 	if ( Freq->name[0] )
@@ -2332,9 +2332,9 @@ char* makeDVBCFreqString( char*Buf, int Size,  DVB_C_FREQ * Freq  )
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	SPRINTF( tmp, sizeof(tmp), "sid:%d ", Freq->sid );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "frq:%ld ", Freq->frequency );
+	SPRINTF( tmp, sizeof(tmp), "frq:%d ", Freq->frequency );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "rate:%ld ", Freq->symbol_rate );
+	SPRINTF( tmp, sizeof(tmp), "rate:%d ", Freq->symbol_rate );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	
 	if ( Freq->modulation != 0xff ) 
@@ -2376,7 +2376,7 @@ char* makeDVBCFreqString( char*Buf, int Size,  DVB_C_FREQ * Freq  )
 		if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	}
 
-	SPRINTF( tmp, sizeof(tmp), "ctrl:%ld ", Freq->ctrl );
+	SPRINTF( tmp, sizeof(tmp), "ctrl:%d ", Freq->ctrl );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	
 	if ( Freq->name[0] )
@@ -2405,9 +2405,9 @@ char* makeDVBSFreqString( char*Buf, int Size,  DVB_S_FREQ * Freq  )
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	SPRINTF( tmp, sizeof(tmp), "sid:%d ", Freq->sid );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "frq:%ld ", Freq->frequency );
+	SPRINTF( tmp, sizeof(tmp), "frq:%d ", Freq->frequency );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
-	SPRINTF( tmp, sizeof(tmp), "rate:%ld ", Freq->symbol_rate );
+	SPRINTF( tmp, sizeof(tmp), "rate:%d ", Freq->symbol_rate );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	
 	if ( Freq->modulation != 0xff ) 
@@ -2479,7 +2479,7 @@ char* makeDVBSFreqString( char*Buf, int Size,  DVB_S_FREQ * Freq  )
 	}
 
 
-	SPRINTF( tmp, sizeof(tmp), "ctrl:%ld ", Freq->ctrl );
+	SPRINTF( tmp, sizeof(tmp), "ctrl:%d ", Freq->ctrl );
 	if ( (int)strlen( tmp )+ (int)strlen( Buf ) + 1 < Size ) strcat( Buf, tmp );
 	
 	if ( Freq->name[0] )
@@ -2655,10 +2655,10 @@ int SaveDVBSPreDefinedFreqTable( CHANNEL_DATA *Channel )
 	return 0;	
 }
 
-int TranslateJWideString( char* buf_out, int buf_out_size, unsigned short* buf_in )
+int TranslateJWideString( char* buf_out, int buf_out_size, uint16_t* buf_in )
 {
 	int len=0, i=0;
-	unsigned short* p = (unsigned short*)buf_in;
+	uint16_t* p = (uint16_t*)buf_in;
 	wchar_t* v;
 	int ret;
 	if ( buf_out == NULL ) return 0;
@@ -2666,7 +2666,7 @@ int TranslateJWideString( char* buf_out, int buf_out_size, unsigned short* buf_i
 
 	len = min( buf_out_size-1, len+1 );
 	v= (wchar_t*)malloc( len*sizeof(wchar_t) );
-	p = (unsigned short*)buf_in;
+	p = (uint16_t*)buf_in;
 	
 #if defined(__APPLE__)
 	for ( i = 0; i<len; i++ ) v[i] = OSReadLittleInt16(p, i * 2);
@@ -2678,7 +2678,7 @@ int TranslateJWideString( char* buf_out, int buf_out_size, unsigned short* buf_i
 	
 	if ( ret == (size_t)-1 )
 	{
-		p = (unsigned short*)buf_in;
+		p = (uint16_t*)buf_in;
 		for ( i = 0; i<len; i++ ) 
 		     buf_out[i]= (char)( 0xff & p[i] ); 
 		buf_out[len] = 0x0;
@@ -2689,10 +2689,10 @@ int TranslateJWideString( char* buf_out, int buf_out_size, unsigned short* buf_i
 
 #if !defined(__APPLE__)
 //translate big endian charactter
-int TranslateJWideString2( char* buf_out, int buf_out_size, unsigned short* buf_in )
+int TranslateJWideString2( char* buf_out, int buf_out_size, uint16_t* buf_in )
 {
 	int len=0, i=0;
-	unsigned short* p = (unsigned short*)buf_in;
+	uint16_t* p = (uint16_t*)buf_in;
 	wchar_t* v;
 	int ret;
 	if ( buf_out == NULL ) return 0;
@@ -2700,7 +2700,7 @@ int TranslateJWideString2( char* buf_out, int buf_out_size, unsigned short* buf_
 
 	len = min( buf_out_size-1, len+1 );
 	v= (wchar_t*)malloc( len*sizeof(wchar_t) );
-	p = (unsigned short*)buf_in;
+	p = (uint16_t*)buf_in;
 	
 	for ( i = 0; i<len; i++ ) v[i] = ( ((p[i]<<8)&0xff00) | p[i] >> 8 ); 
 	
@@ -2709,7 +2709,7 @@ int TranslateJWideString2( char* buf_out, int buf_out_size, unsigned short* buf_
 	
 	if ( ret == (size_t)-1 )
 	{
-		p = (unsigned short*)buf_in;
+		p = (uint16_t*)buf_in;
 		for ( i = 0; i<len; i++ ) 
 		     buf_out[i]= (char)( 0xff & p[i] ); 
 		buf_out[len] = 0x0;
@@ -2752,7 +2752,7 @@ int ConvertIllegalChar2( char* filename )
 }
 
 #include <sys/stat.h>
-unsigned long fileTimeStamp( char* szFileName )
+uint32_t fileTimeStamp( char* szFileName )
 {
 	struct stat st;
 	if ( szFileName == NULL || szFileName[0] == 0 )
